Use jaxtyping annotations for orient transform ndarray types

+++ pycuvslam/apis/track_odometry.py
@@ -6,7 +6,7 @@ import cuvslam
 import numpy as np
 import rerun as rr
 import rerun.blueprint as rrb
-from jaxtyping import Float32, UInt8
+from jaxtyping import Float32, Float64, UInt8
 from scipy.spatial.transform import Rotation
 from simplecv.rerun_log_utils import RerunTyroConfig
 from tqdm.auto import tqdm
@@ -79,7 +79,7 @@ def main(config: TrackOdometryConfig) -> None:
 
     # Tracking loop
     trajectory: list[Float32[np.ndarray, "3"]] = []
-    world_from_rig_matrices: list[Float32[np.ndarray, "4 4"]] = []
+    world_from_rig_matrices: list[Float64[np.ndarray, "4 4"]] = []
     n_cameras: int = len(dataset.cameras)
     rig_from_cam: np.ndarray | None = None
     if config.auto_orient:
